권한 추가

+++ src/main/java/cultureinfo/culture_app/controller/ArticleController.java
@@ -5,24 +5,41 @@ import cultureinfo.culture_app.dto.request.ArticleUpdateDto;
 import cultureinfo.culture_app.dto.response.ArticleDto;
 import cultureinfo.culture_app.dto.response.ArticleSummaryDto;
 import cultureinfo.culture_app.service.ArticleService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.GrantedAuthority;
 import org.springframework.web.bind.annotation.*;
 
+import cultureinfo.culture_app.security.SecurityUtil;
+import cultureinfo.culture_app.repository.MemberRepository;
+import cultureinfo.culture_app.domain.Member;
+import cultureinfo.culture_app.domain.Article;
+import cultureinfo.culture_app.repository.ArticleRepository;
+
+
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/articles")
 public class ArticleController {
     private final ArticleService articleService;
+    private final SecurityUtil securityUtil;
+    private final MemberRepository memberRepository;
+    private final ArticleRepository articleRepository;
 
     //게시글 생성
     @PostMapping
     public ResponseEntity<ArticleDto> createArticle(
-            @RequestParam Long memberId,
             @RequestBody ArticleRequestDto requestDto) {
+
+        //현재 로그인한 사용자의 id를 가져오는 함수
+        Long memberId = securityUtil.getCurrentId();
+
         ArticleDto created = articleService.createArticle(requestDto, memberId);
         return ResponseEntity
                 .status(HttpStatus.CREATED)
@@ -56,6 +73,26 @@ public class ArticleController {
     public ResponseEntity<ArticleDto> updateArticle(
             @PathVariable Long id,
             @RequestBody ArticleUpdateDto requestDto) {
+
+        //현재 로그인한 사용자의 id를 가져오는 함수
+        Long memberId = securityUtil.getCurrentId();
+
+        //가져온 id 정보로 user 정보 가져오기
+        Member member = memberRepository.findById(memberId)
+                        .orElseThrow(() -> new EntityNotFoundException("회원이 존재하지 않습니다."));
+
+        Article article = articleRepository.findById(id)
+                        .orElseThrow(() -> new EntityNotFoundException("글이 존재하지 않습니다."));
+
+        String role = member.getRoles().stream()
+                .map(GrantedAuthority::getAuthority)
+                .collect(Collectors.joining(","));
+        
+        //admin이거나 본인일 때 수정가능하게 설정
+        if(!(article.getMember().getId() == memberId) && role =="ROLE_USER"){
+                throw new AccessDeniedException("본인의 글만 수정할 수 있습니다.");
+        }
+
         ArticleDto updated = articleService.updateArticle(id, requestDto);
         return ResponseEntity.ok(updated);
     }
@@ -63,6 +100,26 @@ public class ArticleController {
     // 게시글 삭제
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteArticle(@PathVariable Long id) {
+
+        //현재 로그인한 사용자의 id를 가져오는 함수
+        Long memberId = securityUtil.getCurrentId();
+
+        //가져온 id 정보로 user 정보 가져오기
+        Member member = memberRepository.findById(memberId)
+                        .orElseThrow(() -> new EntityNotFoundException("회원이 존재하지 않습니다."));
+
+        Article article = articleRepository.findById(id)
+                        .orElseThrow(() -> new EntityNotFoundException("글이 존재하지 않습니다."));
+
+        String role = member.getRoles().stream()
+                .map(GrantedAuthority::getAuthority)
+                .collect(Collectors.joining(","));
+        
+        //admin이거나 본인일 때 삭제가능하게 설정
+        if(!(article.getMember().getId() == memberId) && role =="ROLE_USER"){
+                throw new AccessDeniedException("본인의 글만 수정할 수 있습니다.");
+        }
+
         articleService.deleteArticle(id);
         return ResponseEntity.noContent().build();
     }
